Ark-bot_V5/requirements.txt
discord
google-oauth2-tool
gspread